Class and folders renames
