+++ NSrtm.Core/NSrtm.Core.csproj
@@ -87,11 +87,11 @@
     <Compile Include="Hgt\DataCells\HgtDataCellInvalid.cs" />
     <Compile Include="Pgm\DataDesciption\PgmDataDescriptionExtractor.cs" />
     <Compile Include="Pgm\PgmDataConverter.cs" />
-    <Compile Include="Pgm\PgmGridGraphFactory.cs" />
+    <Compile Include="Pgm\PgmGeoidUndulationGridFactory.cs" />
     <Compile Include="Pgm\DataDesciption\PgmDataDescription.cs" />
-    <Compile Include="Pgm\GridGraph\PgmGridGraphInFile.cs" />
-    <Compile Include="Pgm\GridGraph\PgmGridGraphInMemory.cs" />
-    <Compile Include="Pgm\GridGraph\IPgmGridGraph.cs" />
+    <Compile Include="Pgm\GeoidUndulationGrid\PgmGeoidUndulationGridInFile.cs" />
+    <Compile Include="Pgm\GeoidUndulationGrid\PgmGeoidUndulationGridInMemory.cs" />
+    <Compile Include="Pgm\GeoidUndulationGrid\IPgmGeoidUndulationGrid.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Hgt\HgtElevationProvider.cs" />
   </ItemGroup>
