+++ NSrtm.Core/Pgm/GeoidUndulationGrid/IPgmGeoidUndulationGrid.cs
@@ -1,0 +1,7 @@
+﻿namespace NSrtm.Core.Pgm.GeoidUndulationGrid
+{
+    internal interface IPgmGeoidUndulationGrid
+    {
+        double GetUndulation(double latitude, double longitude);
+    }
+}
