--- NSrtm.Core/Pgm/GridGraph/IPgmGridGraph.cs
@@ -1,7 +1,0 @@
-﻿namespace NSrtm.Core.Pgm.GridGraph
-{
-    internal interface IPgmGridGraph
-    {
-        double GetUndulation(double latitude, double longitude);
-    }
-}
